Update AlphaAnimals_CE_Patch_Race_Gallatross.xml
The creature originally has a 40x health scale which doesn't work well with CE at all, it will invariably bleed to death way before bullets put even a dent to its enormous health pool, rendering it pointless. I'm thinking of making of adding a hediff to the creature which reduces bleeding. but either way, the healthscale needs to be reduced.